generate slugs from collections name with regex substition, relocate helper functions from /hooks.ts to /utils.ts
